requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
Werkzeug==3.0.4
torch>=2.2.0
torchvision>=0.17.0
matplotlib==3.7.1
numpy<2
jupyterlab>3.0
stability-sdk
